Move the TOC code into a separate module file
