+++ js/global.js
@@ -16,6 +16,7 @@ import EnableFlyout from './modules/enable-flyout.js';
 import enableVisibleOnFocus from './modules/enable-visible-on-focus.js';
 import offscreenObserver from './modules/offscreen-observer.js';
 import textZoom from './demos/hero-image-text-resize.js';
+import tableOfContents from './modules/enable-toc.js';
 
 function scrollToEl(el) {
     /*
@@ -80,15 +81,6 @@ function initEnable() {
         },
     );
 
-    // Table of Contents container setup
-    const toc = document.createElement('aside');
-    toc.setAttribute('id', 'toc');
-    const tocHeading = document.createElement('h2');
-    tocHeading.textContent = 'Page Overview';
-    const tocList = document.createElement('ol');
-    tocList.setAttribute('class', 'toc-level-1');
-    let prevHeadingLevel = 0;
-    let tocNode = tocList;
 
 
     pauseAnimControl.init();
@@ -149,55 +141,14 @@ function initEnable() {
                     el.innerHTML = `<a class="heading__deeplink" href="#${el.id}" title="Permalink to ${el.innerText}" aria-label="Permalink to ${el.innerText}">${el.innerHTML}</a>`
                 }
 
-                /**
-                 * Generates a table of contents for your document based on the headings
-                 *  present. Anchors are injected into the document and the
-                 *  entries in the table of contents are linked to them.
-                 */
-                const headingLevel = Number(el.nodeName?.toLowerCase()?.split('h')?.[1] || 0);
-                if (headingLevel === 1) {
-                    prevHeadingLevel = headingLevel;
-                    return;
-                } else if (headingLevel > prevHeadingLevel && prevHeadingLevel !== 1) {
-                    const subList = document.createElement('ul');
-                    subList.setAttribute('class', `toc-level-${headingLevel - 1}`);
-                    tocNode.appendChild(subList);
-                    tocNode = subList;
-                }
-
-                tocNode = tocNode.closest(`.toc-level-${headingLevel - 1}`);
-                
-                const tocItem = document.createElement('li');
-                tocItem.setAttribute('class', `toc-item-${el.tagName.toLowerCase()}`);
-
-                const tocLink = el.id ? document.createElement('a') : document.createElement('p');
-                el.id && tocLink.setAttribute('href', `#${el.id}`);
-                tocLink.textContent = el.textContent;
-            
-                tocItem.appendChild(tocLink);
-                tocNode.appendChild(tocItem);
-                tocNode = tocItem;
-                prevHeadingLevel = headingLevel;
                 return;
             })
         };
 
-    // Add the generated Table of Contents to the top of pages
-    if (location.href.indexOf('index.php') === -1 && location.href.indexOf('faq.php') === -1) {
-        toc.appendChild(tocHeading);
-        toc.appendChild(tocList);
-        const h1 = document.getElementsByTagName('h1')?.[0];
-        const main = document.getElementsByTagName('main')?.[0];
-
-        // Insert the TOC after the H1, or just inside the main element
-        if (h1) {
-            h1.after(toc);
-        } else {
-            main.insertAdjacentElement('afterbegin', toc);
-        }
-    }
 
     focusDeepLink();
+
+    tableOfContents.init(['/index.php', '/faq.php']);
 }
 
 initEnable();
